Unified Ui - incorrect background layer is applied for survey input
MOB-1945
