+++ widgetssdk/src/main/java/com/glia/widgets/view/unifiedui/config/survey/SurveyInputQuestionRemoteConfig.kt
@@ -1,6 +1,5 @@
 package com.glia.widgets.view.unifiedui.config.survey
 
-import com.glia.widgets.view.unifiedui.config.base.LayerRemoteConfig
 import com.glia.widgets.view.unifiedui.config.base.TextRemoteConfig
 import com.glia.widgets.view.unifiedui.theme.survey.SurveyInputQuestionTheme
 import com.google.gson.annotations.SerializedName
@@ -13,15 +12,11 @@ internal data class SurveyInputQuestionRemoteConfig(
     @SerializedName("option")
     val option: OptionButtonRemoteConfig?,
 
-    @SerializedName("background")
-    val background: LayerRemoteConfig?,
-
     @SerializedName("text")
     val textRemoteConfig: TextRemoteConfig?
 ) {
     fun toSurveyInputQuestionTheme(): SurveyInputQuestionTheme = SurveyInputQuestionTheme(
         title = title?.toTextTheme(),
-        background = background?.toLayerTheme(),
         text = textRemoteConfig?.toTextTheme(),
         option = option?.toOptionButtonTheme()
     )
