+++ widgetssdk/src/main/java/com/glia/widgets/view/unifiedui/theme/survey/SurveyInputQuestionTheme.kt
@@ -1,11 +1,9 @@
 package com.glia.widgets.view.unifiedui.theme.survey
 
-import com.glia.widgets.view.unifiedui.theme.base.LayerTheme
 import com.glia.widgets.view.unifiedui.theme.base.TextTheme
 
 internal data class SurveyInputQuestionTheme(
     val title: TextTheme? = null,
     val option: OptionButtonTheme? = null,
-    val background: LayerTheme? = null,
     val text: TextTheme? = null
 )
